resolves build issues stemming from incorrect merge conflic resolution
