+++ provider/provider_config/local_file_store_config.go
@@ -1,0 +1,26 @@
+package provider_config
+
+import (
+	"encoding/json"
+	"fmt"
+)
+
+type LocalFileStoreConfig struct {
+	DirPath string
+}
+
+func (config *LocalFileStoreConfig) Serialize() ([]byte, error) {
+	data, err := json.Marshal(config)
+	if err != nil {
+		panic(err)
+	}
+	return data, nil
+}
+
+func (config *LocalFileStoreConfig) Deserialize(data []byte) error {
+	err := json.Unmarshal(data, config)
+	if err != nil {
+		return fmt.Errorf("deserialize file blob store config: %w", err)
+	}
+	return nil
+}
